+++ Content.Shared/_KS14/OreVent/Drone/SharedOreVentDroneSystem.cs
@@ -1,14 +1,18 @@
 using Content.Shared.Chat;
+using Robust.Shared.Audio;
+using Robust.Shared.Audio.Systems;
+using Robust.Shared.Player;
 using Robust.Shared.Random;
 using Robust.Shared.Spawners;
 
-namespace Content.Shared._KS14.OreVent;
+namespace Content.Shared._KS14.OreVent.Drone;
 
 public abstract class SharedOreVentDroneSystem : EntitySystem
 {
     [Dependency] private readonly IRobustRandom _robustRandom = default!;
     [Dependency] private readonly SharedAppearanceSystem _appearanceSystem = default!;
     [Dependency] private readonly SharedChatSystem _chatSystem = default!;
+    [Dependency] private readonly SharedAudioSystem _audioSystem = default!;
 
     public override void Initialize()
     {
@@ -52,7 +56,10 @@ public abstract class SharedOreVentDroneSystem : EntitySystem
         timedDespawnComponent.Lifetime = 4f;
         AddComp(entity, timedDespawnComponent);
 
-        if (_robustRandom.Prob(0.5f))
-            _chatSystem.TryEmoteWithChat(entity.Owner, "Flip", ignoreActionBlocker: true);
+        if (_robustRandom.Prob(0.25f))
+        {
+            _chatSystem.TryEmoteWithChat(entity.Owner, "Flip", ignoreActionBlocker: true, networkedFilter: Filter.Pvs(entity.Owner, entityManager: EntityManager));
+            _audioSystem.PlayPvs(new SoundPathSpecifier("/Audio/Machines/twobeep.ogg"), entity.Owner);
+        }
     }
 }
