+++ Content.Shared/_KS14/OreVent/OreVentSystem.cs
@@ -1,3 +1,4 @@
+using Content.Shared._KS14.OreVent.Drone;
 using Content.Shared._KS14.ScanDiscoverable.Base;
 using Content.Shared.DoAfter;
 using Content.Shared.Explosion.EntitySystems;
