+++ Resources/Prototypes/_KS14/Entities/Objects/Mining/brm.yml
@@ -1,0 +1,2 @@
+# - type: entity
+#   parent: [ BaseMachinePowered, ConstructibleMachine ]
